medical_app/requirements.txt
streamlit
python-dotenv
bytez
pillow